Refactored the code into react, added assets, configured slideshow to be background
I moved the code from main to vite react and converted plain html, css, and js into jsx and css.

I created a backend in python using Flask.

I reworked the slideshow so instead of it being a rectangle it's become part of the background to make it ascetically pleasing.

I added all of the fonts, icons, and images for the website from main.

What's left to do is make the navigation links accessible and actually have things in them. The major ones are "About Us", "Contact Us", and "Donate"